refactor: Separate and prepare for gatewayEnrollments in DB
